WIP: Efficient AD group matching via wbclient interface
By using the new direct wbclient interface, we get the user's
token with all the SIDs of the groups he is a member-of (inlcuding
nested and primary group), returned as part of NTLM authentication.

Save this information on the request as AD-Group-SID attribute,
so it could be used in policy to match a SID string, like:
if (&config:AD-Group-SID[*] == "S-1-5-21-...")

In addition, register a group-compare function for AD-Group
attribute, which takes an AD group name, resolves its SID via
wbclient interface and looks for a matching AD-Group-SID
on the config list, which allows for:
if (AD-Group == "DOM\Domain Users")

The mapping from group-name to SID can be cached.